requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.5
Django==2.1.5
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
pycparser==2.19
PyJWT==1.7.1
pytz==2018.9
six==1.12.0
